update some feedback from Erik
