+++ components/earlyAccessView/earlyAccess.module.scss
@@ -5,25 +5,37 @@
   right: 50px;
   bottom: 90px;
   z-index: 999;
+  display: flex;
+  justify-content: space-between;
+  left: 50px;
+  width: calc(100vw - 100px);
+  @include mobile {
+    display: unset;
+    justify-content: unset;
+    left: unset;
+    width: unset;
+  }
+  
   @include smallMobile {
     right: 16px;
     bottom: 16px;
+
   }
   .button {
-    width: 196px;
-    height: 94px;
+    width: 166px;
+    height: 64px;
     cursor: pointer;
     display: flex;
     align-items: center;
     justify-content: center;
-    background-image: url('/button/earlyAccess/default.svg');
+    background-image: url('/button/earlyAccess/default.png');
     background-size: 100% 100%;
     border: unset;
     &:hover {
-      background-image: url('/button/earlyAccess/hover.svg');
+      background-image: url('/button/earlyAccess/hover.png');
     }
     &:active {
-      background-image: url('/button/earlyAccess/disable.svg');
+      background-image: url('/button/earlyAccess/disable.png');
     }
   }
   .wrapperForm {
@@ -52,6 +64,9 @@
         cursor: pointer;
         width: 24px;
         height: 24px;
+        &:hover {
+          opacity: 0.6;
+        }
       }
     }
     .emailLabel {
@@ -114,6 +129,18 @@
       font-size: 14px;
       line-height: 214%;
       color: #FFFFFF;
+      &:hover {
+        background-color: #00BA88;
+      }
+    }
+  }
+  .bottomPanel {
+    display: none;
+    @include desktop {
+      display: flex;
+      width: 50px;
+      height: 50px;
+      position: relative;
     }
   }
 }
